backend/requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
django-filter>=23.5

# Database
psycopg2-binary>=2.9

# Authentication
djangorestframework-simplejwt>=5.3

# API Documentation
drf-spectacular>=0.27

# Environment & Config
django-environ>=0.11
python-decouple>=3.8

# Task Queue
celery>=5.3
redis>=5.0
django-celery-beat>=2.5

# File Storage
django-storages>=1.14
boto3>=1.34

# Utilities
Pillow>=10.2
gunicorn>=21.2

# Testing
pytest>=8.0
pytest-django>=4.8
pytest-cov>=4.1
factory-boy>=3.3

# Code Quality
ruff>=0.2
mypy>=1.8
django-stubs>=4.2
